+++ packages/playground-ui/src/ds/components/Columns/column-content.tsx
@@ -1,0 +1,10 @@
+import { cn } from '@/index';
+
+export type ColumnProps = {
+  children?: React.ReactNode;
+  className?: string;
+};
+
+export function ColumnContent({ children, className }: ColumnProps) {
+  return <div className={cn(`grid overflow-y-auto gap-8 content-start`, className)}>{children}</div>;
+}
