+++ packages/playground-ui/src/ds/components/Columns/column-root.tsx
@@ -1,0 +1,25 @@
+import { cn } from '@/index';
+import { WithLabel } from '../Checkbox/checkbox.stories';
+
+export type ColumnProps = {
+  children: React.ReactNode;
+  className?: string;
+  withRightSeparator?: boolean;
+  withLeftSeparator?: boolean;
+};
+
+export function ColumnRoot({ children, className, withLeftSeparator, withRightSeparator }: ColumnProps) {
+  return (
+    <div className="COLUMN flex overflow-y-auto w-full">
+      {withLeftSeparator && <Separator />}
+
+      <div className={cn(`grid gap-6 content-start w-full overflow-y-auto`, className)}>{children}</div>
+
+      {withRightSeparator && <Separator />}
+    </div>
+  );
+}
+
+function Separator() {
+  return <div className={cn('bg-yellow-500 w-[3px] shrink-0 mx-[1.5vw]')}></div>;
+}
