+++ packages/playground-ui/src/ds/components/Columns/column-toolbar.tsx
@@ -1,0 +1,10 @@
+import { cn } from '@/index';
+
+export type ColumnToolbarProps = {
+  children?: React.ReactNode;
+  className?: string;
+};
+
+export function ColumnToolbar({ children, className }: ColumnToolbarProps) {
+  return <div className={cn(`flex items-center justify-end`, className)}>{children}</div>;
+}
