+++ packages/playground-ui/src/ds/components/Columns/index.ts
@@ -1,0 +1,2 @@
+export * from './column';
+export * from './columns';
