+++ packages/playground-ui/src/ds/components/EntryList/types.ts
@@ -1,4 +1,4 @@
-export type Column = {
+export type ColumnType = {
   name: string;
   label: string;
   size: string;
