+++ packages/playground-ui/src/ds/components/ListAndDetails/list-and-details-root.tsx
@@ -5,20 +5,18 @@ import { transitions } from '@/ds/primitives/transitions';
 export type ListAndDetailsRootProps = {
   children: React.ReactNode;
   isDetailsActive?: boolean;
-  isSecondDetailsActive?: boolean; // for future use when we have 3 columns
   className?: string;
 };
 
 export function ListAndDetailsRoot({
   children,
   isDetailsActive = false,
-  isSecondDetailsActive = false,
   className,
 }: ListAndDetailsRootProps): React.JSX.Element {
   return (
     <div
       className={cn(
-        'grid h-full overflow-hidden',
+        'LAD-ROOT grid h-full overflow-hidden',
         transitions.allSlow,
         {
           'grid-cols-1': !isDetailsActive,
