+++ packages/playground-ui/src/index.ts
@@ -81,6 +81,7 @@ export * from './lib/rule-engine';
 
 // DS Components - New
 export * from './ds/components/ListAndDetails';
+export * from './ds/components/Columns';
 
 // Form utilities (AutoForm)
 export * from './lib/form';
